Analyze clean-up  - PartAnalyzeStep, FieldAnalyzeStep, SurfaceAnalyzeStep also INT(KIND=8) as they are compared with iter  - Moved all Poynting vector variables from mesh to analyze and the respective read-in of the parameter CalcPoyntingVecIntegral  - Define parameters description for Poynting vector integral  - Removed obsolete read-in variable PoyntingIntPlaneFactor

+++ src/particles/particle_operations.f90
@@ -183,7 +183,7 @@ IF(PRESENT(BCID)) THEN
   IF(UseNeutralization)THEN
     IF(TRIM(BoundaryName(BCID)).EQ.TRIM(NeutralizationSource))THEN
       ! Add +1 for electrons and -1 for ions
-      NeutralizationBalance = NeutralizationBalance - SIGN(1.0, Species(iSpec)%ChargeIC)
+      NeutralizationBalance = NeutralizationBalance - INT(SIGN(1.0, Species(iSpec)%ChargeIC))
     END IF
   END IF ! UseNeutralization
 END IF ! PRESENT(BCID)
